Move Java function implementations to a sub-module
