+++ sdks/java/extensions/sql/zetasql/src/main/java/org/apache/beam/sdk/extensions/sql/zetasql/SupportedZetaSqlBuiltinFunctions.java
@@ -19,25 +19,17 @@ package org.apache.beam.sdk.extensions.sql.zetasql;
 
 import com.google.zetasql.ZetaSQLFunction.FunctionSignatureId;
 import java.util.List;
-import java.util.Map;
-import org.apache.beam.sdk.annotations.Internal;
-import org.apache.beam.vendor.calcite.v1_20_0.org.apache.calcite.sql.SqlOperator;
-import org.apache.beam.vendor.calcite.v1_20_0.org.apache.calcite.sql.fun.SqlStdOperatorTable;
 import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.collect.ImmutableList;
-import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.collect.ImmutableMap;
 
-/** SqlStdOperatorMappingTable. */
-@Internal
-public class SqlStdOperatorMappingTable {
-
-  /**
-   * Keep this list in sync with
-   * https://github.com/google/zetasql/blob/master/zetasql/public/builtin_function.proto. Uncomment
-   * the corresponding entries to enable parser support to the operators/functions.
-   *
-   * <p>Last synced ZetaSQL release: 2020.06.01
-   */
-  static final List<FunctionSignatureId> ZETASQL_BUILTIN_FUNCTION_ALLOWLIST =
+/**
+ * List of ZetaSQL builtin functions supported by Beam ZetaSQL. Keep this list in sync with
+ * https://github.com/google/zetasql/blob/master/zetasql/public/builtin_function.proto. Uncomment
+ * the corresponding entries to enable parser support to the operators/functions.
+ *
+ * <p>Last synced ZetaSQL release: 2020.06.01
+ */
+class SupportedZetaSqlBuiltinFunctions {
+  static final List<FunctionSignatureId> ALLOWLIST =
       ImmutableList.of(
           FunctionSignatureId.FN_ADD_DOUBLE, // $add
           FunctionSignatureId.FN_ADD_INT64, // $add
@@ -623,82 +615,4 @@ public class SqlStdOperatorMappingTable {
           // FunctionSignatureId.FN_SAFE_ARRAY_AT_OFFSET, // $safe_array_at_offset
           // FunctionSignatureId.FN_SAFE_ARRAY_AT_ORDINAL, // $safe_array_at_ordinal
           );
-
-  // todo: Some of operators defined here are later overridden in ZetaSQLPlannerImpl.
-  // We should remove them from this table and add generic way to provide custom
-  // implementation. (Ex.: timestamp_add)
-  public static final Map<String, SqlOperator> ZETASQL_FUNCTION_TO_CALCITE_SQL_OPERATOR =
-      ImmutableMap.<String, SqlOperator>builder()
-          // grouped window function
-          .put("TUMBLE", SqlStdOperatorTable.TUMBLE)
-          .put("HOP", SqlStdOperatorTable.HOP)
-          .put("SESSION", SqlStdOperatorTable.SESSION)
-
-          // ZetaSQL functions
-          .put("$and", SqlStdOperatorTable.AND)
-          .put("$or", SqlStdOperatorTable.OR)
-          .put("$not", SqlStdOperatorTable.NOT)
-          .put("$equal", SqlStdOperatorTable.EQUALS)
-          .put("$not_equal", SqlStdOperatorTable.NOT_EQUALS)
-          .put("$greater", SqlStdOperatorTable.GREATER_THAN)
-          .put("$greater_or_equal", SqlStdOperatorTable.GREATER_THAN_OR_EQUAL)
-          .put("$less", SqlStdOperatorTable.LESS_THAN)
-          .put("$less_or_equal", SqlStdOperatorTable.LESS_THAN_OR_EQUAL)
-          .put("$like", SqlOperators.LIKE)
-          .put("$is_null", SqlStdOperatorTable.IS_NULL)
-          .put("$is_true", SqlStdOperatorTable.IS_TRUE)
-          .put("$is_false", SqlStdOperatorTable.IS_FALSE)
-          .put("$add", SqlStdOperatorTable.PLUS)
-          .put("$subtract", SqlStdOperatorTable.MINUS)
-          .put("$multiply", SqlStdOperatorTable.MULTIPLY)
-          .put("$unary_minus", SqlStdOperatorTable.UNARY_MINUS)
-          .put("$divide", SqlStdOperatorTable.DIVIDE)
-          .put("concat", SqlOperators.CONCAT)
-          .put("substr", SqlOperators.SUBSTR)
-          .put("trim", SqlOperators.TRIM)
-          .put("replace", SqlOperators.REPLACE)
-          .put("char_length", SqlOperators.CHAR_LENGTH)
-          .put("starts_with", SqlOperators.START_WITHS)
-          .put("ends_with", SqlOperators.ENDS_WITH)
-          .put("ltrim", SqlOperators.LTRIM)
-          .put("rtrim", SqlOperators.RTRIM)
-          .put("reverse", SqlOperators.REVERSE)
-          .put("$count_star", SqlStdOperatorTable.COUNT)
-          .put("max", SqlStdOperatorTable.MAX)
-          .put("min", SqlStdOperatorTable.MIN)
-          .put("avg", SqlStdOperatorTable.AVG)
-          .put("sum", SqlStdOperatorTable.SUM)
-          .put("any_value", SqlStdOperatorTable.ANY_VALUE)
-          .put("count", SqlStdOperatorTable.COUNT)
-          // .put("bit_and", SqlStdOperatorTable.BIT_AND) //JIRA link:
-          // https://issues.apache.org/jira/browse/BEAM-10379
-          .put("string_agg", SqlOperators.STRING_AGG_STRING_FN) // NULL values not supported
-          .put("bit_or", SqlStdOperatorTable.BIT_OR)
-          .put("ceil", SqlStdOperatorTable.CEIL)
-          .put("floor", SqlStdOperatorTable.FLOOR)
-          .put("mod", SqlStdOperatorTable.MOD)
-          .put("timestamp", SqlOperators.TIMESTAMP_OP)
-          .put("$case_no_value", SqlStdOperatorTable.CASE)
-
-          // if operator - IF(cond, pos, neg) can actually be mapped directly to `CASE WHEN cond
-          // THEN pos ELSE neg`
-          .put("if", SqlStdOperatorTable.CASE)
-
-          // $case_no_value specializations
-          // all of these operators can have their operands adjusted to achieve the same thing with
-          // a call to $case_with_value
-          .put("$case_with_value", SqlStdOperatorTable.CASE)
-          .put("coalesce", SqlStdOperatorTable.CASE)
-          .put("ifnull", SqlStdOperatorTable.CASE)
-          .put("nullif", SqlStdOperatorTable.CASE)
-          .build();
-
-  public static final Map<String, SqlOperatorRewriter>
-      ZETASQL_FUNCTION_TO_CALCITE_SQL_OPERATOR_REWRITER =
-          ImmutableMap.<String, SqlOperatorRewriter>builder()
-              .put("$case_with_value", new SqlCaseWithValueOperatorRewriter())
-              .put("coalesce", new SqlCoalesceOperatorRewriter())
-              .put("ifnull", new SqlIfNullOperatorRewriter())
-              .put("nullif", new SqlNullIfOperatorRewriter())
-              .build();
 }
