+++ sdks/java/extensions/sql/zetasql/src/main/java/org/apache/beam/sdk/extensions/sql/zetasql/translation/SqlIfNullOperatorRewriter.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.extensions.sql.zetasql;
+package org.apache.beam.sdk.extensions.sql.zetasql.translation;
 
 import java.util.List;
 import org.apache.beam.vendor.calcite.v1_20_0.org.apache.calcite.rex.RexBuilder;
@@ -26,26 +26,24 @@ import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.base.Preconditio
 import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.collect.ImmutableList;
 
 /**
- * Rewrites NULLIF calls as CASE ($case_no_value) calls.
+ * Rewrites IFNULL calls as CASE ($case_no_value) calls.
  *
- * <p>Turns <code>NULLIF(expression, expression_to_match)</code> into: <code><pre>CASE
- *   WHEN expression == expression_to_match THEN NULL
- *   ELSE expression
+ * <p>Turns <code>IFNULL(expr, null_result)</code> into: <code><pre>CASE
+ *   WHEN expr IS NULL THEN null_result
+ *   ELSE expr
  *   END</pre></code>
  */
-public class SqlNullIfOperatorRewriter implements SqlOperatorRewriter {
+class SqlIfNullOperatorRewriter implements SqlOperatorRewriter {
   @Override
   public RexNode apply(RexBuilder rexBuilder, List<RexNode> operands) {
     Preconditions.checkArgument(
-        operands.size() == 2, "NULLIF should have two arguments in function call.");
+        operands.size() == 2, "IFNULL should have two arguments in function call.");
 
-    SqlOperator op =
-        SqlStdOperatorMappingTable.ZETASQL_FUNCTION_TO_CALCITE_SQL_OPERATOR.get("$case_no_value");
+    SqlOperator op = SqlStdOperatorTable.CASE;
     List<RexNode> newOperands =
         ImmutableList.of(
-            rexBuilder.makeCall(
-                SqlStdOperatorTable.EQUALS, ImmutableList.of(operands.get(0), operands.get(1))),
-            rexBuilder.makeNullLiteral(operands.get(1).getType()),
+            rexBuilder.makeCall(SqlStdOperatorTable.IS_NULL, ImmutableList.of(operands.get(0))),
+            operands.get(1),
             operands.get(0));
 
     return rexBuilder.makeCall(op, newOperands);
